+++ Arsemi-Core/Sensor/SensorData.cs
@@ -5,9 +5,9 @@ namespace Arsemi {
     namespace Sensor {
         [StructLayout(LayoutKind.Sequential, Pack = 1)]
         public struct SensorData() {
-            public float Value = 0;
-            [JsonInclude] public uint ID;
-            [JsonInclude] public uint IntervalMS;
+            public byte Value = 0;
+            [JsonInclude] public byte ID;
+            [JsonInclude] public byte IntervalMS;
         }
     }
 }
